\name{LogDis-package}
\alias{LogDis-package}
\alias{LogDis}
\docType{package}
\title{
\packageTitle{LogDis}
}
\description{
\packageDescription{LogDis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LogDis}
\packageIndices{LogDis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LogDis}

Maintainer: \packageMaintainer{LogDis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
